Refactor tenant payments to use server-side calculation and custom hook
