+++ hooks/use-modal-store.tsx
@@ -4,7 +4,7 @@ import { WasserzaehlerModalData } from '@/types/optimized-betriebskosten';
 import { Tenant, KautionData } from '@/types/Tenant';
 import { Template } from '@/types/template';
 import { ConfirmationDialogVariant } from '@/components/ui/confirmation-dialog';
-import { TenantBentoItem } from '@/components/tenant-payment-bento';
+import { TenantBentoItem } from '@/types/tenant-payment';
 
 // Overview Modal Types
 interface HausWithWohnungen {
@@ -630,7 +630,7 @@ const MODAL_ANIMATION_DURATION = 300; // ms
 export const useModalStore = create<ModalState>((set, get) => {
   let confirmationModalTimeoutId: NodeJS.Timeout | null = null;
 
-  
+
 
   const createCloseHandler = (
     isDirtyFlag: DirtyFlagKey,
@@ -657,9 +657,9 @@ export const useModalStore = create<ModalState>((set, get) => {
 
   return {
     ...createInitialModalState(),
-    openTenantModal: (initialData, wohnungen) => set({ 
-      isTenantModalOpen: true, 
-      tenantInitialData: initialData, 
+    openTenantModal: (initialData, wohnungen) => set({
+      isTenantModalOpen: true,
+      tenantInitialData: initialData,
       tenantModalWohnungen: wohnungen || [],
       isTenantModalDirty: false, // Reset dirty state on open
     }),
@@ -776,7 +776,7 @@ export const useModalStore = create<ModalState>((set, get) => {
 
     // Haus Overview Modal
     openHausOverviewModal: async (hausId: string) => {
-      set({ 
+      set({
         isHausOverviewModalOpen: true,
         hausOverviewLoading: true,
         hausOverviewError: undefined,
@@ -800,24 +800,24 @@ export const useModalStore = create<ModalState>((set, get) => {
 
         // Race between fetch and timeout
         const data = await Promise.race([fetchPromise, timeoutPromise]);
-        
-        set({ 
+
+        set({
           hausOverviewData: data,
-          hausOverviewLoading: false 
+          hausOverviewLoading: false
         });
       } catch (error) {
-        set({ 
+        set({
           hausOverviewError: error instanceof Error ? error.message : 'An error occurred',
-          hausOverviewLoading: false 
+          hausOverviewLoading: false
         });
       }
     },
     refreshHausOverviewData: async () => {
       const state = get();
       if (!state.hausOverviewData?.id) return;
-      
+
       set({ hausOverviewLoading: true, hausOverviewError: undefined });
-      
+
       try {
         const response = await fetch(`/api/haeuser/${state.hausOverviewData.id}/overview`);
         if (!response.ok) {
@@ -826,9 +826,9 @@ export const useModalStore = create<ModalState>((set, get) => {
         const data = await response.json();
         set({ hausOverviewData: data, hausOverviewLoading: false });
       } catch (error) {
-        set({ 
+        set({
           hausOverviewError: error instanceof Error ? error.message : 'An error occurred',
-          hausOverviewLoading: false 
+          hausOverviewLoading: false
         });
       }
     },
@@ -841,7 +841,7 @@ export const useModalStore = create<ModalState>((set, get) => {
 
     // Wohnung Overview Modal
     openWohnungOverviewModal: async (wohnungId: string) => {
-      set({ 
+      set({
         isWohnungOverviewModalOpen: true,
         wohnungOverviewLoading: true,
         wohnungOverviewError: undefined,
@@ -865,24 +865,24 @@ export const useModalStore = create<ModalState>((set, get) => {
 
         // Race between fetch and timeout
         const data = await Promise.race([fetchPromise, timeoutPromise]);
-        
-        set({ 
+
+        set({
           wohnungOverviewData: data,
-          wohnungOverviewLoading: false 
+          wohnungOverviewLoading: false
         });
       } catch (error) {
-        set({ 
+        set({
           wohnungOverviewError: error instanceof Error ? error.message : 'An error occurred',
-          wohnungOverviewLoading: false 
+          wohnungOverviewLoading: false
         });
       }
     },
     refreshWohnungOverviewData: async () => {
       const state = get();
       if (!state.wohnungOverviewData?.id) return;
-      
+
       set({ wohnungOverviewLoading: true, wohnungOverviewError: undefined });
-      
+
       try {
         const response = await fetch(`/api/wohnungen/${state.wohnungOverviewData.id}/overview`);
         if (!response.ok) {
@@ -891,9 +891,9 @@ export const useModalStore = create<ModalState>((set, get) => {
         const data = await response.json();
         set({ wohnungOverviewData: data, wohnungOverviewLoading: false });
       } catch (error) {
-        set({ 
+        set({
           wohnungOverviewError: error instanceof Error ? error.message : 'An error occurred',
-          wohnungOverviewLoading: false 
+          wohnungOverviewLoading: false
         });
       }
     },
@@ -906,7 +906,7 @@ export const useModalStore = create<ModalState>((set, get) => {
 
     // Tenant Payment Overview Modal
     openTenantPaymentOverviewModal: () => {
-      set({ 
+      set({
         isTenantPaymentOverviewModalOpen: true,
         tenantPaymentOverviewLoading: true,
         tenantPaymentOverviewError: undefined,
@@ -922,7 +922,7 @@ export const useModalStore = create<ModalState>((set, get) => {
 
     // Apartment-Tenant Details Modal
     openApartmentTenantDetailsModal: async (apartmentId: string, tenantId?: string) => {
-      set({ 
+      set({
         isApartmentTenantDetailsModalOpen: true,
         apartmentTenantDetailsLoading: true,
         apartmentTenantDetailsError: undefined,
@@ -937,10 +937,10 @@ export const useModalStore = create<ModalState>((set, get) => {
       });
 
       try {
-        const url = tenantId 
+        const url = tenantId
           ? `/api/apartments/${apartmentId}/tenant/${tenantId}/details`
           : `/api/apartments/${apartmentId}/details`;
-        
+
         const fetchPromise = fetch(url).then(async (response) => {
           if (!response.ok) {
             throw new Error('Failed to fetch apartment-tenant details');
@@ -950,29 +950,29 @@ export const useModalStore = create<ModalState>((set, get) => {
 
         // Race between fetch and timeout
         const data = await Promise.race([fetchPromise, timeoutPromise]);
-        
-        set({ 
+
+        set({
           apartmentTenantDetailsData: data,
-          apartmentTenantDetailsLoading: false 
+          apartmentTenantDetailsLoading: false
         });
       } catch (error) {
-        set({ 
+        set({
           apartmentTenantDetailsError: error instanceof Error ? error.message : 'An error occurred',
-          apartmentTenantDetailsLoading: false 
+          apartmentTenantDetailsLoading: false
         });
       }
     },
     refreshApartmentTenantDetailsData: async () => {
       const state = get();
       if (!state.apartmentTenantDetailsData?.apartment?.id) return;
-      
+
       set({ apartmentTenantDetailsLoading: true, apartmentTenantDetailsError: undefined });
-      
+
       try {
         const url = state.apartmentTenantDetailsData.tenant
           ? `/api/apartments/${state.apartmentTenantDetailsData.apartment.id}/tenant/${state.apartmentTenantDetailsData.tenant.id}/details`
           : `/api/apartments/${state.apartmentTenantDetailsData.apartment.id}/details`;
-        
+
         const response = await fetch(url);
         if (!response.ok) {
           throw new Error('Failed to refresh apartment-tenant details');
@@ -980,9 +980,9 @@ export const useModalStore = create<ModalState>((set, get) => {
         const data = await response.json();
         set({ apartmentTenantDetailsData: data, apartmentTenantDetailsLoading: false });
       } catch (error) {
-        set({ 
+        set({
           apartmentTenantDetailsError: error instanceof Error ? error.message : 'An error occurred',
-          apartmentTenantDetailsLoading: false 
+          apartmentTenantDetailsLoading: false
         });
       }
     },
@@ -1055,7 +1055,7 @@ export const useModalStore = create<ModalState>((set, get) => {
       confirmationModalTimeoutId = setTimeout(() => {
         set({ confirmationModalConfig: null });
         confirmationModalTimeoutId = null;
-      }, MODAL_ANIMATION_DURATION); 
+      }, MODAL_ANIMATION_DURATION);
     },
 
     // Folder Delete Confirmation Modal
@@ -1100,7 +1100,7 @@ export const useModalStore = create<ModalState>((set, get) => {
       isTemplateEditorModalOpen: true,
       templateEditorData: {
         template,
-        onSave: onSave || (() => {}),
+        onSave: onSave || (() => { }),
       },
       isTemplateEditorModalDirty: false,
     }),
